Pull common functionality into new base class.
